Introduce `resatectl log seal` to force seal a log
Summary:
By force sealing a log, the cluster controll will take care of
reconfiguring it. It will respect the current cluster configuration
`seen` by the cluster controller.

This process has to be done over the leader cluster controller node. Failure
to do so will return an error.

Once the log is (eventually) sealed the controller will take care of
reconfiguring the log