Metadata HTTP API
Summary:
An interface that exposes Metadata via a REST API.
Is mainly used for testing and not enabled by default

Only enabled with `metadata-api` build flag

Usage:
The API is served by the admin port under `/metadata/<key>`
GET, and HEAD operations always populates the `X-Version` header
if a value is found.

PUT and DELETE operations both require a `X-Precondition` header.
If the precondition is not met, the operation fails with (412 Precondition Failed)

PUT operation requires a header `X-Version` which sets the version
of the new value