Drop `ReplicationStrategy`
Summary:

The idea here is that we can't use `ReplicationStategy`
because it's not location aware and instead we should describe
placement in terms of `ReplicationProperty`.

If placement stragety is not set the system assume placement is done
on all nodes, otherwise the num of copies described by the replication
property is used.

> Right now the placement is still not location aware but this will
most probably change in the future