Introduce `PartitinReplication` and remove `ReplicationStrategy`
Summary:

The idea here is that we can't use `ReplicationStategy`
because it's not location aware and instead we should describe
placement in terms of `ReplicationProperty`.

If partition replication is not set the system assume placement is `Everywhere`,
otherwise the num of copies is limited by the described by the replication
property.

> Right now the placement is still not location aware but this will
most probably change in the future