Fetch metadata via the proxy in restatectl
Summary:
This is mainly to start using metadata proxy (if available)
to fetch the metadata needed for restatectl operation.

These are mainly `NodesConfiguration` and `Logs`.

If the proxy is not available, it will stil do a
majority read for all possible nodes

We will also start to print a warning if restatectl
detects that it's running against a newer version of
the server